Backend/requirements.txt
fastapi
uvicorn
pyyaml
pymodbus
requests